pre-fetch tracks for purchases/sales
