+++ packages/common/src/api/reducer.ts
@@ -4,6 +4,7 @@ import { collectionApiReducer } from './collection'
 import { developerAppsApiReducer } from './developerApps'
 import { favoritesApiReducer } from './favorites'
 import { libraryApiReducer } from './library'
+import { purchasesApiReducer } from './purchases'
 import { relatedArtistsApiReducer } from './relatedArtists'
 import { trackApiReducer } from './track'
 import { trendingApiReducer } from './trending'
@@ -17,5 +18,6 @@ export default combineReducers({
   developerAppsApi: developerAppsApiReducer,
   favoritesApi: favoritesApiReducer,
   trendingApi: trendingApiReducer,
-  libraryApi: libraryApiReducer
+  libraryApi: libraryApiReducer,
+  purchasesApi: purchasesApiReducer
 })
